+++ gtotree/cli/gtt.py
@@ -9,8 +9,7 @@ SUBCOMMAND_MAP = {
     "get-accs-from-ncbi":  "gtotree.utils.ncbi.get_accessions_from_ncbi",
     "hmms":                "gtotree.utils.hmms.gtt_hmms",
     "gen-scg-hmms":        "gtotree.utils.hmms.gen_scg_hmms.gen_scg_hmms_cli",
-    "search-kos":          "gtotree.utils.ko.search_kos_cli",
-    "search-pfams":        "gtotree.utils.pfam.search_pfams_cli",
+    "search-annotations":  "gtotree.utils.target_search.target_search_multi",
     "update-headers":      "gtotree.utils.update_headers.update_headers_cli",
     "midpoint-root-tree":  "gtotree.utils.helper_scripts.gtt_midpoint_root_tree",
     "itol":                "gtotree.cli.itol",
@@ -47,15 +46,11 @@ PROGRAM_GROUPS = [
         ],
     },
     {
-        "title": "Function searching",
+        "title": "KO/Pfam searching",
         "programs": [
             {
-                "name": "search-kos",
-                "desc": "search input genomes for a set of target KOs",
-            },
-            {
-                "name": "search-pfams",
-                "desc": "search input genomes for a set of target Pfams",
+                "name": "search-annotations",
+                "desc": "search input genomes for target KOs and/or Pfams",
             },
         ],
     },